Isolate legacy planned meal compatibility storage
