+++ tests/features/component/today/TodayPage.spec.ts
@@ -3,7 +3,8 @@ import { beforeEach, describe, expect, it } from 'vitest';
 import { getHealthDb } from '$lib/core/db/client';
 import { currentLocalDay } from '$lib/core/domain/time';
 import { saveWorkoutTemplate } from '$lib/features/movement/service';
-import { saveFoodCatalogItem, savePlannedMeal } from '$lib/features/nutrition/service';
+import { savePlannedMeal } from '$lib/features/nutrition/legacy-planned-meal-store';
+import { saveFoodCatalogItem } from '$lib/features/nutrition/service';
 import { ensureWeeklyPlan, savePlanSlot } from '$lib/features/planning/service';
 import TodayPage from '../../../../src/routes/today/+page.svelte';
 import { expectHeading, resetRouteDb, waitForText } from '../../../support/component/routeHarness';
