+++ tests/features/unit/nutrition/controller.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { useTestHealthDb } from '../../../support/unit/testDb';
-import { savePlannedMeal } from '$lib/features/nutrition/service';
+import { savePlannedMeal } from '$lib/features/nutrition/legacy-planned-meal-store';
 import {
   applyNutritionSearchMatches,
   clearNutritionPlannedMeal,
